Cleaned up directories
- Seperated questions and solutions
- Gave every question a unique name
- Put every question in its own .txt file
- Every question has its own folder in "solutions"
