--- issue2sol/readme
@@ -1,1 +1,0 @@
-ADD ISSUE 2 SOLUTIONS in THIS DIRECTORY
