+++ questions/flipped-bits.txt
@@ -11,8 +11,8 @@ Print the number of bits needed to be flipped.
 
 Constraints:
 
-1 ≤ T ≤ 100
-1 ≤ a,b ≤ 1000
+1 <= T <= 100
+1 <= a,b <= 1000
 
 Example:
 
@@ -30,18 +30,3 @@ Explanation:
 A  = 1001001
 B  = 0010101
 No of bits need to flipped = set bit count i.e. 4
-
-
-def countsetbits(b):
-    count=0
-    while(b>0):
-        count=count+(b&1)
-        b=b>>1
-    return count
-    
-t = int(raw_input())
-while t>0:
-    a=map(int,raw_input().split())
-    b=a[0]^a[1]
-    print countsetbits(b)
-    t=t-1
