+++ questions/number-sequence.txt
@@ -1,0 +1,20 @@
+Print a sequence of numbers starting with N, without using loop, in which A[i+1] = A[i] - 5, if A[i]>0, else A[i+1]=A[i] + 5 repeat it until A[i]=N.
+
+Input:
+First line contains number of test cases T. Then following T lines contains an integer N.
+
+Output:
+Single line with pattern .
+
+Constraints:
+0 < N < 10000
+
+Example:
+Input:
+2
+16
+10
+
+Output: 
+16 11 6 1 -4 1 6 11 16
+10 5 0 5 10
