+++ questions/sparse-number.txt
@@ -25,13 +25,3 @@ Output:
 0
 1
 0
-
-n = int(raw_input())
-while(n>0):
-    a = int(raw_input())
-    j = a>>1
-    if(a & j):
-        print '0'
-    else:
-        print '1'
-    n=n-1
