+++ questions/sum-of-digits.txt
@@ -11,9 +11,9 @@ Calculate the sum of digits of N.
 
 Constraints:
 
-1 ≤ T ≤ 30
+1 <= T <= 30
 
-1 ≤ N ≤ 1000
+1 <= N <= 1000
 
 Example:
 
@@ -25,15 +25,3 @@ Input
 Output
 6
 9
-
-def toto(a):
-    if(a==0):
-        return 0
-    else:
-        return (a%10)+toto(a/10)
-        
-t = int(raw_input())
-while(t>0):
-    n=int(raw_input())
-    print toto(int(n))
-    t=t-1
