+++ solutions/check-binary-tree/check-binary-tree.py
