+++ solutions/other-solutions/issue1sol/issue1sol.py
