+++ solutions/other-solutions/issue1sol/sol1.py
